requirements.txt
certifi==2018.1.18
chardet==3.0.4
dropbox==8.7.0
idna==2.6
pkg-resources==0.0.0
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
